Refactor components and actions into separate files
