+++ src/actions/updateProfile.ts
@@ -2,8 +2,6 @@
 
 import { UpdateProfileValues, updateProfileSchema } from "@/lib/validation";
 
-// To learn more about server actions, watch my YT tutorial: https://www.youtube.com/watch?v=XD5FpbVpWzk
-
 export async function updateProfile(values: UpdateProfileValues) {
   // TODO: Get the currently authenticated user
