+++ src/components/NavBar.tsx
@@ -4,8 +4,8 @@ export default function NavBar() {
   // TODO: Show the currently logged-in user
 
   return (
-    <header className="sticky top-0 bg-background px-3 shadow-sm">
-      <nav className="mx-auto flex h-14 w-full max-w-7xl items-center justify-between gap-3">
+    <header className="sticky top-0 px-3 shadow-sm bg-background">
+      <nav className="flex items-center justify-between w-full gap-3 mx-auto h-14 max-w-7xl">
         <Link href="/" className="font-bold">
           Next-Auth v5 Tutorial
         </Link>
